Move autodriving code into a new file
